requirements.txt
python-gitlab==3.3.0
requests==2.27.1

